+++ src/components/IssueList/IssueList.tsx
@@ -1,14 +1,15 @@
 import IssueItem from './IssueItem/IssueItem';
 import { useIssues } from '../../context/IssueContext';
-import { useEffect, useRef } from 'react';
 import * as S from './IssueList.style';
 import AdBox from '../common/list/AdBox';
 import Loading from '../common/list/Loading';
 import HasNoMore from '../common/list/HasNoMore';
 import Error from '../common/list/Error';
+import useIntersect from 'hook/useIntersect';
 
 const IssueList = () => {
   const { handleGetIssues, issues, loading, hasMore } = useIssues();
+  const observerRef = useIntersect(() => hasMore && handleGetIssues());
 
   const adBoxProps = {
     alt: '광고',
@@ -16,34 +17,6 @@ const IssueList = () => {
     linkTo: 'https://www.wanted.co.kr/',
     width: '478px',
     height: '100px',
-  };
-
-  const observerRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const observerOptions = {
-      root: null,
-      rootMargin: '0px',
-      threshold: 1.0,
-    };
-    const observer = new IntersectionObserver(handleObserver, observerOptions);
-    if (observerRef.current) {
-      observer.observe(observerRef.current);
-    }
-    return () => {
-      if (observerRef.current) {
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-        observer.unobserve(observerRef.current);
-      }
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  const handleObserver = (entries: IntersectionObserverEntry[]) => {
-    const target = entries[0];
-    if (target.isIntersecting) {
-      handleGetIssues();
-    }
   };
 
   return (
